Provide single launcher and unit tests
Restructure puzzle code to be testable.
